separate tests for rdm_plot
